refactor: Move chain-storage-paths.js to avoid a module cycle
